requirements.txt
# Core
fastapi==0.115.2
uvicorn==0.34.2
sqlalchemy==2.0.41
alembic==1.13.1
pydantic==2.11.5
pydantic_core==2.33.2

# Segurança / Autenticação
passlib[bcrypt]==1.7.4
python-jose==3.4.0
bcrypt==4.3.0
python-multipart==0.0.20

# Validações
email-validator==2.2.0
annotated-types==0.7.0

# Testes
pytest==8.3.5
pytest-asyncio==0.26.0

# HTTP cliente / servidor
httpx==0.28.1
httpcore==1.0.9
starlette==0.46.2
anyio==4.9.0
sniffio==1.3.1
h11==0.16.0
httptools==0.6.4
websockets==15.0.1
watchfiles==1.0.5

# Utilidades
python-dotenv==1.1.0
typing-extensions==4.13.2
typing-inspection==0.4.1
greenlet==3.2.2
colorama==0.4.6
packaging==25.0
click==8.2.1
iniconfig==2.1.0
pluggy==1.6.0

# Segurança (opcional para JWT / criptografia)
rsa==4.9.1
ecdsa==0.19.1
pyasn1==0.4.8

# Certificados
certifi==2025.4.26
idna==3.10
six==1.17.0
